Shortened modules to spring-cloud-aws-v3-secrets-manager and spring-cloud-starter-aws-v3-secrets-manager. Also moved all autoconfig classes to spring-cloud-aws-v3-secrets-manager module. Hence, pom.xml for spring-cloud-starter-aws-v3-secrets-manager now includes just the dependency for spring-cloud-aws-v3-secrets-manager.
