+++ v3/spring-cloud-aws-v3-secrets-manager/src/main/java/io/awspring/cloud/v3/secretsmanager/AwsSecretsManagerProperties.java
