+++ v3/spring-cloud-aws-v3-secrets-manager/src/main/resources/META-INF/spring.factories
@@ -1,0 +1,7 @@
+# ConfigData Location Resolvers
+org.springframework.boot.context.config.ConfigDataLocationResolver=\
+io.awspring.cloud.v3.secretsmanager.autoconfigure.AwsSecretsManagerConfigDataLocationResolver
+
+# ConfigData Loaders
+org.springframework.boot.context.config.ConfigDataLoader=\
+io.awspring.cloud.v3.secretsmanager.autoconfigure.AwsSecretsManagerConfigDataLoader
