+++ v3/spring-cloud-aws-v3-secrets-manager/src/test/java/io/awspring/cloud/v3/secretsmanager/AwsSecretsManagerPropertiesTest.java
