--- v3/spring-cloud-starter-aws-v3-secrets-manager-config/pom.xml
@@ -1,32 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<project xmlns="http://maven.apache.org/POM/4.0.0"
-		 xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-		 xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
-	<parent>
-		<artifactId>v3</artifactId>
-		<groupId>io.awspring.cloud</groupId>
-		<version>2.3.0-RC2</version>
-	</parent>
-	<modelVersion>4.0.0</modelVersion>
-
-	<artifactId>spring-cloud-starter-aws-v3-secrets-manager-config</artifactId>
-	<name>Spring Cloud AWS V3 Secrets Manager Configuration Starter</name>
-	<description>Spring Cloud AWS V3 Secrets Manager Configuration Starter</description>
-
-	<dependencies>
-		<!-- note that we don't depend on spring-cloud-aws-autoconfigure: that module brings in
-	 		 spring-cloud-aws-context and unwanted auto-configuration when you just want to use
-	 		 the Secrets Manager configuration support.
-	 		 We simply include our own auto-configuration for Spring Cloud in this starter instead. -->
-		<dependency>
-			<groupId>io.awspring.cloud</groupId>
-			<artifactId>spring-cloud-aws-v3-secrets-manager-config</artifactId>
-		</dependency>
-
-		<dependency>
-			<groupId>io.awspring.cloud</groupId>
-			<artifactId>spring-cloud-aws-v3-autoconfigure</artifactId>
-		</dependency>
-
-	</dependencies>
-</project>
