--- v3/spring-cloud-starter-aws-v3-secrets-manager-config/src/main/resources/META-INF/spring.factories
@@ -1,7 +1,0 @@
-# ConfigData Location Resolvers
-org.springframework.boot.context.config.ConfigDataLocationResolver=\
-io.awspring.cloud.v3.autoconfigure.secretsmanager.AwsSecretsManagerConfigDataLocationResolver
-
-# ConfigData Loaders
-org.springframework.boot.context.config.ConfigDataLoader=\
-io.awspring.cloud.v3.autoconfigure.secretsmanager.AwsSecretsManagerConfigDataLoader
