+++ v3/spring-cloud-starter-aws-v3-secrets-manager/pom.xml
@@ -1,0 +1,24 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<project xmlns="http://maven.apache.org/POM/4.0.0"
+		 xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+		 xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
+	<parent>
+		<artifactId>v3</artifactId>
+		<groupId>io.awspring.cloud</groupId>
+		<version>2.3.0-RC2</version>
+	</parent>
+	<modelVersion>4.0.0</modelVersion>
+
+	<artifactId>spring-cloud-starter-aws-v3-secrets-manager</artifactId>
+	<name>Spring Cloud AWS V3 Secrets Manager Starter</name>
+	<description>Spring Cloud AWS V3 Secrets Manager Starter</description>
+
+	<dependencies>
+
+		<dependency>
+			<groupId>io.awspring.cloud</groupId>
+			<artifactId>spring-cloud-aws-v3-secrets-manager</artifactId>
+		</dependency>
+
+	</dependencies>
+</project>
